+++ include/ck_tile/ops/gemm/pipeline/tile_gemm_traits.hpp
@@ -11,8 +11,8 @@ namespace ck_tile {
 template <bool kPadM_,
           bool kPadN_,
           bool kPadK_,
-          typename ALayout_,
-          typename BLayout_,
+          typename AsLayout_,
+          typename BsLayout_,
           typename CLayout_,
           index_t NumWaveGroups_ = 1>
 struct TileGemmTraits
@@ -24,9 +24,9 @@ struct TileGemmTraits
     // TODO this can't be hardcoded here! Should be in policy!
     static constexpr int _VectorSize = 16;
 
-    using ALayout = ALayout_;
-    using BLayout = BLayout_;
-    using CLayout = CLayout_;
+    using AsLayout = AsLayout_;
+    using BsLayout = BsLayout_;
+    using CLayout  = CLayout_;
 
     static constexpr bool TransposeC            = false;
     static constexpr bool UseStructuredSparsity = false;
@@ -37,8 +37,8 @@ template <bool kPadM_,
           bool kPadN_,
           bool kPadK_,
           bool DoubleSmemBuffer_,
-          typename ALayout_,
-          typename BLayout_,
+          typename AsLayout_,
+          typename BsLayout_,
           typename CLayout_,
           bool TransposeC_                   = false,
           bool UseStructuredSparsity_        = false,
@@ -54,9 +54,9 @@ struct TileGemmUniversalTraits
     static constexpr int _VectorSize       = 16;
     static constexpr bool DoubleSmemBuffer = DoubleSmemBuffer_;
 
-    using ALayout = ALayout_;
-    using BLayout = BLayout_;
-    using CLayout = CLayout_;
+    using AsLayout = AsLayout_;
+    using BsLayout = BsLayout_;
+    using CLayout  = CLayout_;
 
     static constexpr bool TransposeC                   = TransposeC_;
     static constexpr bool UseStructuredSparsity        = UseStructuredSparsity_;
@@ -70,8 +70,8 @@ template <bool kPadM_,
           bool kPadN_,
           bool kPadK_,
           bool DoubleSmemBuffer_,
-          typename ALayout_,
-          typename BLayout_,
+          typename AsLayout_,
+          typename BsLayout_,
           typename CLayout_,
           bool TransposeC_            = false,
           bool UseStructuredSparsity_ = false>
@@ -79,8 +79,8 @@ using PersistentTileGemmUniversalTraits = TileGemmUniversalTraits<kPadM_,
                                                                   kPadN_,
                                                                   kPadK_,
                                                                   DoubleSmemBuffer_,
-                                                                  ALayout_,
-                                                                  BLayout_,
+                                                                  AsLayout_,
+                                                                  BsLayout_,
                                                                   CLayout_,
                                                                   TransposeC_,
                                                                   UseStructuredSparsity_,